+++ sniffer/network/engine.py
@@ -3,7 +3,8 @@ import struct
 from pwn import hexdump
 from ..utils.constants import ETH_P_IP
 from .analyzer import PacketAnalyzer
-from ..exceptions.network import SLLUnsupportedError, UninterestingPacketException
+from ..exceptions.network import SLLUnsupportedError,\
+    UninterestingPacketException, UnsupportedVersionException
 
 
 class SnifferEngine:
@@ -11,13 +12,16 @@ class SnifferEngine:
 
     def __init__(self, interface: str):
         self.interface = interface
-        self.socket = socket.socket(socket.AF_PACKET,
+        self.socket = socket.socket(socket.AF_INET,
                                     socket.SOCK_RAW,
-                                    socket.htons(ETH_P_IP))
+                                    socket.IPPROTO_TCP)
+                                    # socket.htons(ETH_P_IP))
 
-        if self.interface != SnifferEngine.NET_INTERFACE_ANY:
+        # if self.interface != SnifferEngine.NET_INTERFACE_ANY:
             # Attach to network interface
-            self.socket.bind((self.interface, 0))
+        # self.socket.bind((self.interface, 0))
+        self.socket.bind(('10.0.3.15', 0))
+        # self.socket.setsockopt(socket.IPPROTO_IP, socket.IP_HDRINCL, 1)
 
     def sniff(self):
         while True:
@@ -33,8 +37,10 @@ class SnifferEngine:
                 print(analyzer.get_source_port())
                 print(analyzer.get_dest_port())
                 print(analyzer.get_content())
-            # except SLLUnsupportedError:
-            #     continue
+            except SLLUnsupportedError:
+                continue
+            except UnsupportedVersionException:
+                continue
             # except struct.error:
             #     continue
             except UninterestingPacketException: